requirements.txt
slack-sdk==3.36.0
fastapi==0.116.1
uvicorn[standard]==0.30.6
python-multipart==0.0.20
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.9.2
openai==1.100.1
Pillow==11.3.0
dropbox==12.0.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.140.0
requests==2.32.3
pytz==2024.1
aiohttp==3.12.15
hijri_converter==2.3.2.post1
redis==5.0.1